clippy: apply cargo clippy --fix across workspace (101 files)
Auto-applied: or_fun_call, assigning_clones, unnecessary_unwrap, derivable_impls,
clone_on_ref_ptr, ptr_as_ptr/ref_as_ptr/borrow_as_ptr/ptr_cast_constness,
vec_init_then_push, map_clone, implicit_clone.

TarballStream.rs reverted: autofix stripped deliberate &(...) wrappers
guarding dangerous_implicit_autorefs.

+++ src/ast/expr.rs
@@ -386,7 +386,7 @@ impl Expr {
         }
 
         if let Some(idx) = bun_core::index_of_any(name, b"[.") {
-            let idx = idx as usize;
+            let idx = idx;
             match name[idx] {
                 b'[' => {
                     let end_idx = bun_core::index_of_char(name, b']')? as usize;
